+++ aofl-js-packages/i18n-mixin/modules/i18n-mixin.js
@@ -88,7 +88,12 @@ export default (superClass) => {
      */
     async __(id, str) {
       const translated = await this.i18n.__(id, str);
-      const templateParts = html([translated]);
+      const templateParts = aoflDevtools.showI18nIds? html([
+        `<span id="${id}" class="aofl-i18n-string" title="${id}" style="position: relative;">${translated}
+          <span style="font-size: 12px; color: #000; position: absolute; left: 0; bottom: 0; transform: translateY(60%); white-space: pre; background: yellow; line-height: 1; padding: 0 2px;">${id.replace('<', '&lt;').replace('>', '&gt;')}</span>
+        </span>`
+      ]): html([translated]);
+
       return this.getCachedTemplate(id, templateParts);
     }
 
@@ -120,7 +125,12 @@ export default (superClass) => {
      */
     async _c(id, str, ...args) {
       const translated = await this.i18n._c(id, str, ...args);
-      const templateParts = html([translated]);
+      const templateParts = aoflDevtools.showI18nIds? html([
+        `<span id="${id}" class="aofl-i18n-string" title="${id}" style="position: relative;">${translated}
+          <span style="font-size: 12px; color: #000; position: absolute; left: 0; bottom: 0; transform: translateY(60%); white-space: pre; background: yellow; line-height: 1; padding: 0 2px;">${id.replace('<', '&lt;').replace('>', '&gt;')}</span>
+        </span>`
+      ]): html([translated]);
+
       return this.getCachedTemplate(id, templateParts);
     }